+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## v4.0-dev - Unreleased
+## v4.0 - 2026-01-22
 
 This release drops support of Python 3.11 and below. The minimum required Python
 version is 3.12.